refactor: moved repositories to separated directories
- removed repositories from services folder
- created Repositories and Repositories/Interfaces directories
- separated repository and interface files from repositories
