+++ API/Repositories/Interfaces/ITokenRepository.cs
@@ -1,0 +1,8 @@
+using API.Models;
+
+namespace API.Repositories.Interfaces;
+
+public interface ITokenRepository
+{
+    public Task AddToken(Token token);
+}
